+++ src/main/java/org/realityforge/gwt/eventsource/client/html5/Html5EventSource.java
@@ -55,9 +55,20 @@ public class Html5EventSource
     _eventSource = EventSourceImpl.create( this, url, withCredentials );
   }
 
+  @Nonnull
   @Override
   public final ReadyState getReadyState()
   {
+    if ( null == _eventSource )
+    {
+      return ReadyState.CLOSED;
+    }
+    else
+    {
+      return ReadyState.values()[ _eventSource.getReadyState() ];
+    }
+  }
+
   @Nonnull
   @Override
   public final String getURL()